design: css 수정 및 html 태그 수정
