+++ src/components/Cart/CartOrderInfo.tsx
@@ -1,17 +1,17 @@
 import { css } from '@emotion/react';
 import { useRecoilValue } from 'recoil';
 
-import Receipt from '../Receipt';
-
-import { INFO_ICON } from '@/assets/images';
+import { INFO_ICON } from '@assets/images';
 import { orderTotalPriceState, deliveryPriceState } from '@recoil/cartItems/selectors';
 
-export default function CartFooterSection() {
+import Receipt from '@components/Receipt';
+
+export default function CartOrderInfo() {
   const totalPrice = useRecoilValue(orderTotalPriceState);
   const deliveryPrice = useRecoilValue(deliveryPriceState);
 
   return (
-    <section css={cartFooterSection}>
+    <div css={cartOrderInfoContainer}>
       <div css={freeDeliveryGuideWrapper}>
         <span css={infoIcon}>
           <img src={INFO_ICON} alt="info icon" />
@@ -19,16 +19,17 @@ export default function CartFooterSection() {
         {/* TODO: 100,000 상수 분리 */}
         <span css={freeDeliveryGuide}>총 주문 금액이 100,000원 이상일 경우 무료 배송됩니다.</span>
       </div>
+
       <div css={borderTopWrapper}>
         <Receipt title="주문 금액" price={totalPrice} />
         <Receipt title="배송비" price={deliveryPrice} />
         <Receipt title="총 결제 금액" price={totalPrice + deliveryPrice} />
       </div>
-    </section>
+    </div>
   );
 }
 
-const cartFooterSection = css`
+const cartOrderInfoContainer = css`
   display: flex;
   flex-direction: column;
   gap: 12px;
