+++ src/pages/CartConfirmPage.tsx
@@ -3,9 +3,9 @@ import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 
-import { fetchCartItems } from '@/apis/cartItem';
-import { cartItemsState } from '@/recoil/cartItems/atoms';
+import { fetchCartItems } from '@apis/cartItem';
 import { BACK_ARROW } from '@assets/images';
+import { cartItemsState } from '@recoil/cartItems/atoms';
 import {
   productTypesCountState,
   purchaseTotalPriceState,
@@ -41,9 +41,7 @@ export default function CartConfirmPage() {
         </button>
       </Header>
       <div css={container}>
-        <div css={titleWrapper}>
-          <h2 css={title}>주문 확인</h2>
-        </div>
+        <h2 css={title}>주문 확인</h2>
         <div css={orderInfoContainer}>
           총 {productTypesCount}종류의 상품 {totalQuantity}개를 주문합니다. <br />
           최종 결제 금액을 확인해주세요.
@@ -68,16 +66,12 @@ const container = css`
   gap: 24px;
 `;
 
-const titleWrapper = css`
-  display: flex;
-  align-items: center;
-  height: 35px;
-  gap: 24px;
-`;
-
 const title = css`
+  height: 35px;
+
   font-size: 24px;
   font-weight: 700;
+  line-height: 35px;
 `;
 
 const orderInfoContainer = css`
